--- src/main/java/duke/Parser.java
@@ -1,7 +1,0 @@
-package duke;
-
-public class Parser {
-    // deals with making sense of the user command
-    // the errors
-
-}
