--- src/main/java/duke/Storage.java
@@ -1,6 +1,0 @@
-package duke;
-
-public class Storage {
-    // deals with loading tasks from the file and saving tasks in the file
-    // ReadFromFile and WriteToFile
-}
