--- src/main/java/duke/Ui.java
@@ -1,7 +1,0 @@
-package duke;
-
-public class Ui {
-    // deals with interactions with the user
-    // get user input -- the if loop
-
-}
