+++ src/main/java/duke/task/Task.java
@@ -9,22 +9,37 @@ public class Task {
         this.isDone = false;
     }
 
+    /**
+     *
+     */
     public String getDescription() {
         return description;
     }
 
+    /**
+     *
+     */
     public void markAsDone() {
         isDone = true;
     }
 
+    /**
+     *
+     */
     public String getStatusIcon() {
         return (isDone ? "\u2713" : "\u2718"); // return tick or X symbols
     }
 
+    /**
+     *
+     */
     public String toString() {
         return "[" + getStatusIcon() + "] " + description;
     }
 
+    /**
+     *
+     */
     public String toFileString() {
         return " " + (isDone ? 1 : 0) + " " + description;
     }
